Create abstract alpinelinux builders
