+++ src/abstract_builders/alpinelinux_rootfs_model.py
@@ -1,0 +1,48 @@
+from pydantic import BaseModel, Field, ConfigDict, StringConstraints
+from typing_extensions import Annotated
+from typing import Optional
+
+from abstract_builders.block_model import Block_Model, Block_Project_Model, Block_Project_Model_Default_Fields
+from abstract_builders.file_system_model import File_System_Installable_Build_Time_Item_Model, File_System_User_Model
+
+
+class AlpineLinux_RootFS_Dependencies_Model(BaseModel):
+    model_config = ConfigDict(extra="forbid", strict=True)
+
+    kernel: str
+    devicetree: Optional[str] = None
+    vivado: Optional[str] = None
+
+
+class AlpineLinux_RootFS_Block_Project_Model(Block_Project_Model):
+    model_config = ConfigDict(extra="forbid", strict=True)
+
+    repositories: list[Annotated[str, StringConstraints(pattern=r"^http")]] = Field(
+        default=["https://dl-cdn.alpinelinux.org/alpine/v3.22/main"],
+        description="List of URLs to Alpine Linux repositories",
+    )
+    addl_pkgs: Optional[list[str]] = Field(default=None, description="List of additional apk packages to be installed")
+    addl_ext_pkgs: Optional[list[str]] = Field(
+        default=None, description="List of additional packages to be installed from external *.apk files"
+    )
+    build_time_fs_layer: Optional[list[File_System_Installable_Build_Time_Item_Model]] = Field(
+        default=None,
+        description="List of files and directories to be installed that were generated by other blocks at build time",
+    )
+    users: list[File_System_User_Model] = Field(default=None, description="List of users to be added")
+    add_build_info: bool = Block_Project_Model_Default_Fields.add_build_info
+    dependencies: AlpineLinux_RootFS_Dependencies_Model = Block_Project_Model_Default_Fields.dependencies
+
+
+class AlpineLinux_RootFS_Block_Model(Block_Model):
+    model_config = ConfigDict(extra="forbid", strict=True)
+
+    project: AlpineLinux_RootFS_Block_Project_Model
+
+
+class AlpineLinux_RootFS_Blocks_Model(BaseModel):
+    model_config = ConfigDict(extra="ignore")
+
+    rootfs: AlpineLinux_RootFS_Block_Model = Field(
+        default=..., description="Configuration of the Alpine Linux root file system block"
+    )
